+++ examples/identity.rb
@@ -20,7 +20,7 @@ module Node
     class Parser
       include CallableTree::Node::Internal
 
-      def match?(input, **options)
+      def match?(input, **_options)
         File.extname(input) == '.json'
       end
 
@@ -31,7 +31,7 @@ module Node
         end
       end
 
-      def terminate?(_output, **)
+      def terminate?(_output, *_inputs, **_options)
         true
       end
     end
@@ -47,11 +47,11 @@ module Node
         Identity.new(klass: super, type: @type)
       end
 
-      def match?(input, **options)
+      def match?(input, **_options)
         !!input[@type.to_s]
       end
 
-      def call(input, **options)
+      def call(input, **_options)
         input[@type.to_s]
           .map { |element| [element['name'], element['emoji']] }
           .to_h
@@ -63,7 +63,7 @@ module Node
     class Parser
       include CallableTree::Node::Internal
 
-      def match?(input, **options)
+      def match?(input, **_options)
         File.extname(input) == '.xml'
       end
 
@@ -73,7 +73,7 @@ module Node
         end
       end
 
-      def terminate?(_output, **)
+      def terminate?(_output, *_inputs, **_options)
         true
       end
     end
@@ -89,11 +89,11 @@ module Node
         Identity.new(klass: super, type: @type)
       end
 
-      def match?(input, **options)
+      def match?(input, **_options)
         !input.get_elements("//#{@type}").empty?
       end
 
-      def call(input, **options)
+      def call(input, **_options)
         input
           .get_elements("//#{@type}")
           .first