+++ spec/callable_tree/node/hooks/call_spec.rb
@@ -6,8 +6,8 @@ RSpec.describe CallableTree::Node::Hooks::Call do
       include CallableTree::Node::External
       prepend CallableTree::Node::Hooks::Call
 
-      def call(input, **)
-        input.reverse
+      def call(input, a, b, x:, y:)
+        "#{a}#{x}#{input}#{y}#{b}".reverse
       end
     end
   end
@@ -15,61 +15,61 @@ RSpec.describe CallableTree::Node::Hooks::Call do
   let(:node) { HooksCallSpec::Reverser.new }
 
   describe '#before_call' do
-    subject { node.call(input, **options) }
+    subject { node.call(*inputs, **options) }
 
-    let(:input) { 'abc' }
-    let(:options) { { foo: '*', bar: '!' } }
+    let(:inputs) { %w[foobar ( )] }
+    let(:options) { { x: '[', y: ']' } }
 
     before do
       node
-        .before_call { |input, foo:, bar:| input.ljust(5, foo) }
-        .before_call { |input, foo:, bar:| input.ljust(7, bar) }
+        .before_call { |input, *, x:, y:, **| "#{x}#{input}#{y}" }
+        .before_call { |input, a, b, *, **| "#{a}#{input}#{b}" }
     end
 
-    it { is_expected.to eq '!!**cba' }
+    it { is_expected.to eq ')])]raboof[([(' }
   end
 
   describe '#around_call' do
-    subject { node.call(input, **options) }
+    subject { node.call(*inputs, **options) }
 
-    let(:input) { 'abc' }
-    let(:options) { { foo: '*', bar: '!' } }
+    let(:inputs) { %w[foobar ( )] }
+    let(:options) { { x: '[', y: ']' } }
 
     before do
       node
-        .around_call do |_input, foo:, bar:, &block|
-          output = block.call
-          "#{foo}: #{output}"
-        end
-        .around_call do |input, foo:, bar:, &block|
-          output = block.call
-          "#{bar}: #{input}, #{output}"
-        end
+        .around_call do |*, x:, y:, &block|
+        output = block.call
+        "#{x}#{output}#{y}"
+      end
+        .around_call do |input, a, b, *, **, &block|
+        output = block.call
+        "#{a}#{input}#{b} -> #{a}#{output}#{b}"
+      end
     end
 
-    it { is_expected.to eq '!: abc, *: cba' }
+    it { is_expected.to eq '(foobar) -> ([)]raboof[(])' }
   end
 
   describe '#after_call' do
-    subject { node.call(input, **options) }
+    subject { node.call(*inputs, **options) }
 
-    let(:input) { 'abc' }
-    let(:options) { { foo: '*', bar: '!' } }
+    let(:inputs) { %w[foobar ( )] }
+    let(:options) { { x: '[', y: ']' } }
 
     before do
       node
-        .after_call { |output, foo:, bar:| output.ljust(5, foo) }
-        .after_call { |output, foo:, bar:| output.ljust(7, bar) }
+        .after_call { |output, *, x:, y:| "#{x}#{output}#{y}" }
+        .after_call { |output, *, x:, y:| "#{x}#{output}#{y}" }
     end
 
-    it { is_expected.to eq 'cba**!!' }
+    it { is_expected.to eq '[[)]raboof[(]]' }
   end
 
   describe '#clone' do
     subject { node.clone }
 
-    let(:before_callback) { proc { |input, **| input } }
-    let(:around_callback) { proc { |_input, **, &block| block.call } }
+    let(:before_callback) { proc { |input, *, **| input } }
+    let(:around_callback) { proc { |_input, *, **, &block| block.call } }
     let(:after_callback) { proc { |output, **| output } }
 
     before do
@@ -80,9 +80,11 @@ RSpec.describe CallableTree::Node::Hooks::Call do
     end
 
     it 'should generate new array' do
-      expect(subject.before_callbacks.object_id).not_to eq node.before_callbacks.object_id
-      expect(subject.around_callbacks.object_id).not_to eq node.around_callbacks.object_id
-      expect(subject.after_callbacks.object_id).not_to eq node.after_callbacks.object_id
+      expect(subject.before_callbacks).not_to be node.before_callbacks
+      expect(subject.around_callbacks).not_to be node.around_callbacks
+      expect(subject.after_callbacks).not_to be node.after_callbacks
     end
   end
+
+  # TODO: add call spec
 end